Strip 'declare -f' prefixes in `list` function
Because IFS is now set to only newlines and spaces, the array created in
`list` doesn't split on spaces anymore, so the full 'declare -f
function' line is treated as a single array element.

In order to account for this change, awk can be used to just get the
last word in each element, which is the function name. The if statement
therefore only needs to check against strings beginning with an
underscore.